chore: move stream and deadletter queues to their own file
